refactor: move and rename LandingPagesHeader
