+++ src/components/SharedComponents/LayoutLandingPagesHeader/index.tsx
@@ -1,0 +1,2 @@
+export { default } from "./LayoutLandingPagesHeader";
+export type { LayoutLandingPagesHeaderProps } from "./LayoutLandingPagesHeader";
